Tests: Add FormBlockEditorCest
